refactor -> Score 처리 부분 ScoreBoard class로 구분해서 처리
